fix: changed int to float on input

+++ main.py
@@ -13,7 +13,7 @@ try:
     while True:
         
         name = str(input("Text:\t"))
-        scale = int(input("Scale:\t"))
+        scale = float(input("Scale:\t"))
         screen.display_text([[name, scale, font]])
 except KeyboardInterrupt:
     screen.epd2in13_V4.epdconfig.module_exit(cleanup=True)
